cambios

+++ src/Actividades/Encuesta.java
@@ -13,7 +13,7 @@ public class Encuesta extends Actividad {
 	private String tipo;
 
 	public Encuesta(String actividadID, String descripcion, String objetivo, int nivelDificultad, int duracionEsperada,
-			boolean esObligatoria, Date fechaLimite, String resenas, double calificacion, int resultado, List<String> actividadesPrevias,
+			boolean esObligatoria, Date fechaLimite, String resenas, double calificacion, double resultado, List<String> actividadesPrevias,
 			List<String> actividadesSeguimiento, List<String> preguntas) {
 		super(actividadID, descripcion, objetivo, nivelDificultad, duracionEsperada, esObligatoria, fechaLimite, resenas,
 				actividadesPrevias, actividadesSeguimiento, ENCUESTA);
