+++ src/Actividades/Examen.java
@@ -14,7 +14,7 @@ public class Examen extends Actividad{
 	
 
 	public Examen(String actividadID, String descripcion, String objetivo, int nivelDificultad, int duracionEsperada,
-			boolean esObligatoria, Date fechaLimite, String resenas, double calificacion, int resultado, List<String> actividadesPrevias,
+			boolean esObligatoria, Date fechaLimite, String resenas, double calificacion, double resultado, List<String> actividadesPrevias,
 			 List<String> actividadesSeguimiento, List<Pregunta> preguntas, double calificacionMinima) {
 
 		super(actividadID, descripcion, objetivo, nivelDificultad, duracionEsperada, esObligatoria, fechaLimite, resenas,
