+++ src/Actividades/Tarea.java
@@ -15,7 +15,7 @@ public class Tarea extends Actividad {
 	private Collection<List<String>> opciones;
 
 	public Tarea(String actividadID, String descripcion, String objetivo, int nivelDificultad, int duracionEsperada,
-			boolean esObligatoria, Date fechaLimite, String resenas, double calificacion, int resultado,List<String> actividadesPrevias,  
+			boolean esObligatoria, Date fechaLimite, String resenas, double calificacion,double resultado,List<String> actividadesPrevias,  
 			List<String> actividadesSeguimiento, List<Pregunta> preguntas, String instrucciones, String estado) {
 		super(actividadID, descripcion, objetivo, nivelDificultad, duracionEsperada, esObligatoria, fechaLimite, resenas,
 				actividadesPrevias, actividadesSeguimiento, TAREA);
